+++ tests/unit/Rules/NotTest.php
@@ -29,11 +29,6 @@ class NotTest extends \PHPUnit_Framework_TestCase
         $this->assertTrue($not->assert($input));
     }
 
-    public function testShortcutNot()
-    {
-        $this->assertTrue(Validator::intVal()->not()->assert('afg'));
-    }
-
     /**
      * @dataProvider providerForInvalidNot
      * @expectedException Respect\Validation\Exceptions\ValidationException
@@ -42,14 +37,6 @@ class NotTest extends \PHPUnit_Framework_TestCase
     {
         $not = new Not($v);
         $this->assertFalse($not->assert($input));
-    }
-
-    /**
-     * @expectedException Respect\Validation\Exceptions\ValidationException
-     */
-    public function testShortcutNotNotHaha()
-    {
-        $this->assertFalse(Validator::intVal()->not()->assert(10));
     }
 
     public function providerForValidNot()