+++ CHANGELOG.md
@@ -38,6 +38,8 @@ All notable changes of the Respect\Validation releases are documented in this fi
 - On exceptions, do not display parent message then rule has only one child (#407)
 - On exceptions, improve `Object` conversion to string (#399)
 - On exceptions, improve conversion of all values by using JSON (#399)
+- Remove `addOr()` shortcut (#444)
+- Remove `not()` shortcut (#444)
 - Remove identical checking from "Equals" rule (#442)
 - Rename rule "Arr" to "ArrayVal"
 - Rename rule "Bool" to "BoolType" (#426)